% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSKAT_helper.R
\name{inner.CSKAT}
\alias{inner.CSKAT}
\title{Inner Function for CSKAT, Correlated Sequence Kernel Association Test}
\usage{
inner.CSKAT(lmer.obj, K)
}
\arguments{
\item{lmer.obj}{A fitted lme4 object (model under H0)}

\item{K}{the kernel matrix, which quantifies the similarities between samples}
}
\value{
\describe{
  \item{p.value}{association p-value}
  \item{Q.adj}{adjusted score statistic}
}
}
\description{
Small-sample SKAT for correlated (continuous) data ('c' stands for 'correlated'). Computes the adjusted score statistic and p-value.
}
\references{
Zhan X, et al. (2018) A small-sample kernel association test for correlated data with application to microbiome association studies.  
Genet Epidemiol., submitted.
}
