\name{throat.meta}
\alias{throat.meta}
\docType{data}
\title{
Meta data of the throat microbiome samples
}
\description{
This data set includes samples from the microbiome of the nasopharynx and oropharynx on each side of the body. It were generated to study the effect of smoking on the microbiota of the upper respiratory tract in 60 individuals, 28 smokers and 32 nonsmokers.
}
\usage{data("throat.meta")}
\format{
  A data frame with 60 observations on 16 variables.
}

\source{
Charlson ES, Chen J, Custers-Allen R, Bittinger K, Li H, et al. (2010) Disordered Microbial Communities in the Upper Respiratory Tract of Cigarette Smokers. PLoS ONE 5(12): e15216.
}
\references{
R package "GUniFrac"
}
\examples{
data(throat.meta)
}
\keyword{datasets}
