% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/throat.otu.tab.R
\docType{data}
\name{throat.otu.tab}
\alias{throat.otu.tab}
\title{Simulated OTU data for microbiome regression analysis}
\format{
60 rows and 856 columns, where rows are patients and columns are OTUs
}
\usage{
data(throat.otu.tab)
}
\description{
Simulated code can be seen in ?KRV
Corresponding metadata is stored in "throat.meta"
}
\keyword{datasets}
