% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/throat.tree.R
\docType{data}
\name{throat.tree}
\alias{throat.tree}
\title{Simulated rooted phylogenetic tree}
\format{
Phylogenetic tree with 856 tips and 855 internal nodes
}
\usage{
data(throat.tree)
}
\description{
Simulation code can be seen in ?KRV
}
\details{
Corresponding OTU matrix stored in "throat.otu.tab"
See the GUniFrac package for more details
}
\keyword{datasets}
