\name{throat.otu.tab}
\alias{throat.otu.tab}
\docType{data}
\title{
OTU count table from 16S sequencing of the throat microbiome samples
}
\description{
This data set contains 60 subjects with 28 smokers and 32 nonsmokers. Microbiome data were collected from right and left nasopharynx and oropharynx region to form an OTU table with 856 OTUs. 
}
\usage{data("throat.otu.tab")}
\format{
  A data frame with 60 observations on 856 variables.
}
\source{
Charlson ES, Chen J, Custers-Allen R, Bittinger K, Li H, et al. (2010) Disordered Microbial Communities in the Upper Respiratory Tract of Cigarette Smokers. PLoS ONE 5(12): e15216.
}
\references{
R package "GUniFrac"
}
\examples{
data(throat.otu.tab)
}
\keyword{datasets}
