\name{throat.tree}
\alias{throat.tree}
\docType{data}
\title{
UPGMA tree of the OTUs from 16S sequencing of the throat microbiome samples
}
\description{
The OTU tree is constructed using UPGMA on the K80 distance matrix of the OTUs. It is a rooted tree of class "phylo".
}
\usage{data("throat.tree")}
\format{
  List of 4 data frames.
}
\source{
Charlson ES, Chen J, Custers-Allen R, Bittinger K, Li H, et al. (2010) Disordered Microbial Communities in the Upper Respiratory Tract of Cigarette Smokers. PLoS ONE 5(12): e15216.
}
\references{
R package "GUniFrac"
}
\examples{
data(throat.tree)
}
\keyword{datasets}
